+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,5 +1,9 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.Pin;
+import bowling.domain.Pins;
+import bowling.domain.Score;
+import bowling.domain.ScoreSymbol;
 import java.util.List;
 import org.springframework.util.ObjectUtils;
