+++ src/main/java/bowling/domain/frame/Frames.java
@@ -1,5 +1,7 @@
-package bowling.domain;
+package bowling.domain.frame;
 
+import bowling.domain.Round;
+import bowling.domain.Score;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
