+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.frame;
+
+import bowling.domain.ScoreSymbol;
 
 public class NormalFrame extends Frame {
