+++ src/test/java/bowling/domain/FramesTest.java
@@ -2,6 +2,7 @@ package bowling.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import bowling.domain.frame.Frames;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
